feat: add NeuronModel base class
This base class will implement transformers PreTrainedModel methods
that are not implemented in optimum PreTrainedModel base class.
