+++ docs/source/package_reference/modeling.mdx
@@ -18,12 +18,12 @@ limitations under the License.
 
 ## Generic model classes
 
-### NeuronBaseModel
+### NeuronTracedModel
 
-The `NeuronBaseModel` class is available for instantiating a base Neuron model without a specific head.
+The `NeuronTracedModel` class is available for instantiating a base Neuron model without a specific head.
 It is used as the base class for all tasks but text generation.
 
-[[autodoc]] modeling_base.NeuronBaseModel
+[[autodoc]] modeling_traced.NeuronTracedModel
 
 ### NeuronDecoderModel
