+++ optimum/neuron/modeling_diffusion.py
@@ -39,7 +39,7 @@ from ..exporters.neuron import (
 from ..exporters.neuron.model_configs import *  # noqa: F403
 from ..exporters.tasks import TasksManager
 from ..utils import is_diffusers_available
-from .modeling_base import NeuronBaseModel
+from .modeling_traced import NeuronTracedModel
 from .utils import (
     DIFFUSION_MODEL_TEXT_ENCODER_2_NAME,
     DIFFUSION_MODEL_TEXT_ENCODER_NAME,
@@ -101,7 +101,7 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class NeuronStableDiffusionPipelineBase(NeuronBaseModel):
+class NeuronStableDiffusionPipelineBase(NeuronTracedModel):
     auto_model_class = StableDiffusionPipeline
     library_name = "diffusers"
     base_model_prefix = "neuron_model"
@@ -319,7 +319,7 @@ class NeuronStableDiffusionPipelineBase(NeuronBaseModel):
             submodels.pop("unet")
             for submodel_name, submodel_path in submodels.items():
                 if submodel_path is not None and submodel_path.is_file():
-                    submodels[submodel_name] = NeuronBaseModel.load_model(submodel_path)
+                    submodels[submodel_name] = NeuronTracedModel.load_model(submodel_path)
                 else:
                     submodels[submodel_name] = None
             submodels["unet"] = torch_neuronx.DataParallel(
@@ -331,7 +331,7 @@ class NeuronStableDiffusionPipelineBase(NeuronBaseModel):
             logger.info("Loading the pipeline without any data parallelism...")
             for submodel_name, submodel_path in submodels.items():
                 if submodel_path is not None and submodel_path.is_file():
-                    submodels[submodel_name] = NeuronBaseModel.load_model(submodel_path)
+                    submodels[submodel_name] = NeuronTracedModel.load_model(submodel_path)
         else:
             raise ValueError("You need to pass `data_parallel_mode` to define Neuron Core allocation.")
 
@@ -820,7 +820,7 @@ class _NeuronDiffusionModelPart:
     def __init__(
         self,
         model: torch.jit._script.ScriptModule,
-        parent_model: NeuronBaseModel,
+        parent_model: NeuronTracedModel,
         config: Optional[Union[DiffusersPretrainedConfig, PretrainedConfig]] = None,
         neuron_config: Optional["NeuronDefaultConfig"] = None,
         model_type: str = "unet",
@@ -845,7 +845,7 @@ class NeuronModelTextEncoder(_NeuronDiffusionModelPart):
     def __init__(
         self,
         model: torch.jit._script.ScriptModule,
-        parent_model: NeuronBaseModel,
+        parent_model: NeuronTracedModel,
         config: Optional[DiffusersPretrainedConfig] = None,
         neuron_config: Optional[Dict[str, str]] = None,
     ):
@@ -882,7 +882,7 @@ class NeuronModelUnet(_NeuronDiffusionModelPart):
     def __init__(
         self,
         model: torch.jit._script.ScriptModule,
-        parent_model: NeuronBaseModel,
+        parent_model: NeuronTracedModel,
         config: Optional[DiffusersPretrainedConfig] = None,
         neuron_config: Optional[Dict[str, str]] = None,
     ):
@@ -918,7 +918,7 @@ class NeuronModelVaeEncoder(_NeuronDiffusionModelPart):
     def __init__(
         self,
         model: torch.jit._script.ScriptModule,
-        parent_model: NeuronBaseModel,
+        parent_model: NeuronTracedModel,
         config: Optional[DiffusersPretrainedConfig] = None,
         neuron_config: Optional[Dict[str, str]] = None,
     ):
@@ -935,7 +935,7 @@ class NeuronModelVaeDecoder(_NeuronDiffusionModelPart):
     def __init__(
         self,
         model: torch.jit._script.ScriptModule,
-        parent_model: NeuronBaseModel,
+        parent_model: NeuronTracedModel,
         config: Optional[DiffusersPretrainedConfig] = None,
         neuron_config: Optional[Dict[str, str]] = None,
     ):
